added style and search bar but have yet to implement search functionality

+++ app/src/main/java/com/example/pokemonapp/data/local/entity/Pokemon.kt
@@ -10,6 +10,6 @@ data class Pokemon(
     val id: Int,
     val name: String,
     val height: Int,
-    val weight: Int,
+    val weight: Long,
     val sprites: String
 )
